Split the agent stub into one file per implementation
